// core/src/data-table/main/foundation/table.api.ts
export interface CellStyles extends Styleable {
	/**
	 * Whether the body cell is secondary info.
	 */
	useSecondaryColor?: boolean;

	/**
	 * The hidden text will be placed at the secondary cell.
	 *
	 * @default placed at the first secondary cell of the row with content:
	 * - English: "Withdrawn"
	 * - German: "Löschen (wiederherstellbar)"
	 */
	secondaryCellTitle?: string;
}